* Clarified some naming between documents/entities:    - A Document is the structure contains Id & Rev. It is a couch specific structure that identifies a record in couch.    - An Entity is a domain object that we happen to be responsible for persisting in couch.
That couch doesn't have a term 'entity' is the whole point. Thats what WE provide: a layer between a specific dbms (couch) and domain entities.

No functionality changed, it was simply that I was calling the generic argument for TDocument (its now TEntity), which was very confusing since the Document structure is also used as argument and return values.

git-svn-id: http://relax-net.googlecode.com/svn/trunk@12 ff391d20-badf-11de-869a-473eb4776397